separate BlogController into BlogRevision and BlogArticle
