+++ routes/web.php
@@ -20,7 +20,6 @@ $router->get('/', function () use ($router) {
 $router->post('/admin/login', ['uses'=>'AdminAuthController@authenticate', 'middleware'=>'throttle:5,1']); // throttled 5 requests/1 min
 $router->get('/admin/user', ['uses'=>'AdminAuthController@user_info', 'middleware'=>'auth:admin']);
 
-$router->get('/blog/revisions/{id}', ['uses' => 'BlogController@get_revision']);
-$router->post('/blog/revisions/', ['uses' => 'BlogController@post_revision']);
-$router->get('/blog/articles/{id}', ['uses' => 'BlogController@get_article']);
+$router->get('/blog/revisions/{id}', ['uses' => 'BlogRevisionController@get_revision']);
+$router->get('/blog/articles/{id}', ['uses' => 'BlogArticleController@get_article']);
 $router->options('{path:.*}', function(){}); // any path
